backend/requirements.txt
Django==4.1.7
psycopg2==2.9.5
django-cockroachdb==4.1.1
Django-ddd==0.1.10
djangorestframework==3.14.0

pytest==7.2.2
pytest-django==4.5.2
ipdb==0.13.13
